Move deploys to its own package
