+++ src/test/java/com/kinbiko/bugsnagmavenplugin/deploys/BugsnagDeployMojoTest.java
@@ -1,4 +1,4 @@
-package com.kinbiko.bugsnagmavenplugin;
+package com.kinbiko.bugsnagmavenplugin.deploys;
 
 import com.mashape.unirest.http.exceptions.UnirestException;
 import org.junit.Test;
